requirements.txt
img2pdf==0.6.0
qq-botpy==1.2.1
jmcomic==2.5.32